[v7r1] Several fixes for scripts

+++ WorkloadManagementSystem/scripts/dirac-admin-show-task-queues.py
@@ -7,9 +7,9 @@
    Show details of currently active Task Queues
 """
 from __future__ import print_function
+
 __RCSID__ = "$Id$"
 
-from past.builtins import long
 import sys
 
 from DIRAC import S_OK, gLogger
@@ -33,7 +33,7 @@ taskQueueID = 0
 
 def setTaskQueueID(optVal):
   global taskQueueID
-  taskQueueID = long(optVal)
+  taskQueueID = int(optVal)
   return S_OK()
